Fix difference in generated testcase vs Raspberry Pi.
The type of vector<sometype>.size() varies between my laptop and the
Raspberry Pi, so rnd.next(vector<sometype>.size()) calls a different overload
on each platform.

Cast to int instead - now the generated testcases are binary equal
between laptop and Raspberry Pi, but I now need to come up with new
seeds for all testcases all over again XD

+++ my-own-problems/shared/random-choosable-set.h
@@ -69,7 +69,7 @@ class RandomChooseableSet
         TPtr chooseRandom() const
         {
             assert(!empty());
-            return m_vec[rnd.next(m_vec.size())];
+            return m_vec[rnd.next(static_cast<int>(m_vec.size()))];
         }
     private:
         std::map<TPtr, int> m_indexFor;
